+++ test/state/CMakeLists.txt
@@ -9,8 +9,7 @@ target_include_directories(evmone-state PRIVATE ${evmone_private_include_dir})
 target_sources(
     evmone-state PRIVATE
     account.hpp
-    blob_schedule.hpp
-    blob_schedule.cpp
+    blob_params.hpp
     block.hpp
     block.cpp
     bloom_filter.hpp
