+++ test/state/block.cpp
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "block.hpp"
-#include "blob_schedule.hpp"
 #include "rlp.hpp"
 #include "transaction.hpp"
