+++ test/statetest/statetest_runner.cpp
@@ -26,7 +26,7 @@ void run_state_test(const StateTransitionTest& test, evmc::VM& vm, bool trace_su
             const auto& expected = cases[case_index];
             const auto tx = test.multi_tx.get(expected.indexes);
             auto state = test.pre_state;
-            const auto blob_params = state::get_blob_params(rev, test.blob_schedule);
+            const auto blob_params = get_blob_params(rev, test.blob_schedule);
 
             const auto res = test::transition(state, block, test.block_hashes, tx, rev, vm,
                 block.gas_limit, static_cast<int64_t>(state::max_blob_gas_per_block(blob_params)));
