+++ test/unittests/statetest_loader_block_info_test.cpp
@@ -20,7 +20,7 @@ TEST(statetest_loader, block_info)
             "withdrawals": []
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -49,7 +49,7 @@ TEST(statetest_loader, block_info_hex)
         "parentUncleHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"
     })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -78,7 +78,7 @@ TEST(statetest_loader, block_info_dec)
         "parentUncleHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"
     })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -105,7 +105,7 @@ TEST(statetest_loader, block_info_0_current_difficulty)
         "parentUncleHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"
     })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -132,7 +132,7 @@ TEST(statetest_loader, block_info_0_parent_difficulty)
         "parentUncleHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"
     })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -156,7 +156,7 @@ TEST(statetest_loader, block_info_0_random)
             "withdrawals": []
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -192,7 +192,7 @@ TEST(statetest_loader, block_info_withdrawals)
             ]
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -231,7 +231,7 @@ TEST(statetest_loader, block_info_ommers)
             "withdrawals": []
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -263,7 +263,7 @@ TEST(statetest_loader, block_info_parent_blob_gas)
             "parentBlobGasUsed": "0x60000"
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -288,7 +288,7 @@ TEST(statetest_loader, block_info_current_blob_gas)
             "currentExcessBlobGas": "2"
         })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.coinbase, 0x1111111111111111111111111111111111111111_address);
     EXPECT_EQ(bi.prev_randao, 0x00_bytes32);
@@ -309,7 +309,7 @@ TEST(statetest_loader, block_info_parent_beacon_block_root)
         "parentBeaconBlockRoot": "0xbeac045007"
     })";
 
-    const auto blob_params = state::get_blob_params(EVMC_CANCUN);
+    const auto blob_params = test::get_blob_params(EVMC_CANCUN);
     const auto bi = test::from_json_with_rev(json::json::parse(input), EVMC_CANCUN, blob_params);
     EXPECT_EQ(bi.parent_beacon_block_root, 0xbeac045007_bytes32);
 }
