+++ test/utils/blob_schedule.cpp
@@ -1,11 +1,10 @@
 #include "blob_schedule.hpp"
-#include "../utils/utils.hpp"
-#include <cstdint>
+#include "utils.hpp"
 #include <string_view>
 
-namespace evmone::state
+namespace evmone::test
 {
-BlobParams get_blob_params(evmc_revision rev)
+state::BlobParams get_blob_params(evmc_revision rev)
 {
     if (rev == EVMC_PRAGUE || rev == EVMC_EXPERIMENTAL)
         return {6, 9, 5007716};
@@ -16,12 +15,12 @@ BlobParams get_blob_params(evmc_revision rev)
         return {3, 6, 3338477};
 }
 
-BlobParams get_blob_params(evmc_revision rev, const BlobSchedule& blob_schedule)
+state::BlobParams get_blob_params(evmc_revision rev, const BlobSchedule& blob_schedule)
 {
     return get_blob_params(evmc::to_string(rev), blob_schedule, 0);
 }
 
-BlobParams get_blob_params(
+state::BlobParams get_blob_params(
     std::string_view network, const BlobSchedule& blob_schedule, int64_t timestamp)
 {
     std::string fork;
@@ -40,7 +39,7 @@ BlobParams get_blob_params(
     if (const auto it = blob_schedule.find(fork); it != blob_schedule.end())
         return it->second;
     else
-        return get_blob_params(test::to_rev_schedule(network).get_revision(timestamp));
+        return get_blob_params(to_rev_schedule(network).get_revision(timestamp));
 }
 
-}  // namespace evmone::state
+}  // namespace evmone::test
